feat(BreadcrumbEllipsis): move the component into the Breadcrumbs
